refactor(telemetry-generator): Upgrade eslint config to "recommended" and fix violations in code (#21076)
This package was using the `minimal-deprecated` configuration, which is
scheduled for deletion. Upgrades the package's config to the
`recommended` base and fixes violations in the code.

This PR disables some of the `recommended` rules due to the sheer number
of violations. These can and should be removed and fixed in the future.
For now, the primary goal here is to remove usages of the deprecated
config.


[AB#2900](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/2900)

---------

Co-authored-by: Alex Villarreal <716334+alexvy86@users.noreply.github.com>