refactor(getKeys): Upgrade eslint config to "recommended" and fix violations in code (#21078)
This package was using the `minimal-deprecated` configuration, which is
scheduled for deletion. Upgrades the package's config to the
`recommended` base and fixes violations in the code.

This PR disables some of the `recommended` rules that are
TypeScript-specific. Longer term, we should create a better pattern for
JavaScript packages. For now, the primary goal here is to remove usages
of the deprecated config.

AB#8051