refactor(container-runtime): Tighten up exposure of ContainerRuntime's logger (#20839)
## Description

ContainerRuntime.logger is currently public, though ideally, it should
not be, and it uses an internal type instead of the preferable base
type. When sharing the logger, the standard practice should be to pass
around the base logger, allowing each consumer to create a ChildLogger
from it.

## Changes

* Rename ContainerRuntime.logger to ContainerRuntime.baseLogger, change
its type to ITelemetryBaseLogger, and set its access level to private.

* Adjust the dependencies that receive the logger to accept the
ITelemetryBaseLogger type and utilize createChildLogger to create a
child logger as needed (this operation will be a no-op if the logger is
already a child logger).

* Modify the following classes in Fluid Framework (FF) so they no longer
directly access the logger from the runtime but instead receive their
own base logger via constructor parameters:

BlobManager
Summarizer
DataStoreContext

* DataStoreContext currently makes the same logger available through its
public API. This may be unnecessary—as indicated by its lack of use in
the FF codebase—especially since IFluidDataStoreRuntime already exposes
a logger. This exposure could potentially be removed to streamline the
API and enhance encapsulation.