Make the new summary retry logic the default and remove old logic (#20851)
https://github.com/microsoft/FluidFramework/pull/16885 added a new retry
mechanism in case of summarization failures. This was rolled out with
A/B testing and the experiment was successful. This PR makes the new
retry mechanism default and removes the old one.

**Notes**
- Added retry for these two additional scenarios - When a summary op or
summary ack is not received within the timeout, retry. This is because
these failures are mostly transient and a retry may fix them.
- Removed `refreshLatestAck` summary option because its not used
anymore. A lot of the code changes are because of this.


[AB#7605](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/7605)