Better telemetry for disconnect events: record pendingClientId (#21093)
When connection does not move far enough (does not get to "Connected" state), we do not have good mechanism to pair "establishing connection" events with "disconnect" events (unless complex join queries are involved), as clientId can't be used for such pairing. We need pendingClientId for that, but it's reset on disconnect before telemetry event fires.

This change moves the time when we reset it, such that it gets into telemetry.