[GC] Remove the option to disable GC mark phase (#21101)
GC is enabled by default and has been this way for quite a while.
Specifically, the mark phase which marks unreferenced objects as such in
the summary is always enabled. This does not effect the container's
content or its usage in any way meaning that it doesn't put any
restrictions on content retrievel and it doesn't delete anything. These
are done by the other two stages of GC - tombstone and sweep. These can
still be controlled via settings.

This PR removes the options to turn GC mark phase off. It will always be
enabled and cannot be turned off:
- `gcAllowed` option was used to control the mark phase roll-out. The
roll out was completed couple of years ago and has been stable. This
option is now removed.
- `disableGC` option was to provide ability to turn it off in case
things went wrong. This was never used (except in tests) and is mow
removed.
- The mark phase can still be turned off in tests via the feature flag
`Fluid.GarbageCollection.Test.RunGC`. This used to be called
`Fluid.GarbageCollection.RunGC` and is renamed in this PR to include
`Test`.