fix(build-tools): fluid-build-tasks-eslint deps (#21089)
getDefaultTscTaskDependencies was ineffective and is now removed. eslint
with ts processing does not require build first. Instead of requiring
that tsc is run for the projects listed in lint config, depend on the
things the respective tsc tasks depend on.

Once policy changes are picked up by client group ^tsc dependencies
listed for eslint in examples/data-objects can be removed.