Offline: Add negative tests to illustrate gap in capability around identifying ops submitted in multiple incarnations of the same container (#21095)
We need to handle the case where a container is
serialized, and then multiple incarnations of the same container
resubmit the same ops. Those resubmitted ops will all be on unique
clientIds, so the existing op ID logic (which just identifies "local"
ops as those matching the current clientId) will not recognize them as
duplicates.

This PR adds 3 test cases illustrating this gap.