[Service load test] Force some connections to be write mode for Tinylicious (#21136)
We are seeing a large number of "OpRoundtripTime" errors for Tinylicious
runs, largely caused by replaying ops at the beginning of the run.
Forcing some of these 120 clients to be in write mode from the start
should help alleviate some of the load.

After this change is merged, I will be monitoring the CI runs to see if
the error rate improves. If not, I'll revert this change.

[AB#7819](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/7819)