Remove several types from `@public` scope (#21142)
## Description

Due to my previous changes to make IFluidDataStoreRuntime non-public,
several types no longer need to be public. This moves more of them to
alpha.

Also includes a couple of `IAudience` related tweaks to make the types
better convey what I I think is supported as far as method binding and
modification.

## Breaking Changes

The following types have been moved from `@public` to `@alpha`:

-   `IFluidSerializer`
-   `ISharedObjectEvents`
-   `IChannelServices`
-   `IChannelStorageService`
-   `IDeltaConnection`
-   `IDeltaHandler`

These should not be needed by users of the declarative API, which is
what `@public` is targeting.