Upload manifest file for release branch (#21143)
The manifest file is uploaded by date and build number for main branch.
This change extends the functionality for release branches too.
 
For release branches, the manifest file is uploaded based on the build
number, rather than the date. This approach prevents it from overriding
the one uploaded from the main branch.
 
Once this PR is merged in main, I can port the change to release
branches.