fix(presence-tracker): rename TinyliciousMember.userName to TinyliciousMember.name  (#21111)
We recently renamed the AzureMember.userName to AzureMember.name to
establish uniform naming across odsp-client and azure-client.

This change needs to be reflected in the presence tracker example since
it uses audience to display focus status and mouse position of all
connected clients. The example right now is essentially just a blank
screen.