Kkachmar/remove session start metric (#21125)
This change removes the SessionStartMetric from Scribe and Deli. The
reasons behind this change are:
1. There has been a bug where the size of the checkpoints is too large,
causing checkpoints to fail.
2. A likely cause of this is the `successfullyStartedLambdas` data
object, which can be very large, and is used in the
`SessionStartMetric`.
3. When checkpoints fail, the `successfullyStartedLambdas` value in the
checkpoint could be incorrect, which may lead to invalid metrics.
4. With updated logs and metrics, we don't need this metric or the
validation of `successfullyStartedLambdas`, as we have logging around
session start behaviors such as the `restoreFromCheckpoint` metric and
`RunService` metric, as well as logging for creating document partitions
for document lambdas like Scribe and Deli.