Add an in-memory and default implementation of detached blob storage (#21144)
Based on feedback from partners, the current story around blobs is
overly complicated, and hard to work with. The existing detached blob
storage interface is difficult to work with as it applies to the whole
loader, and not a specific container, additionally it is decoupled from
the serialization flow which make coordination difficult, lastly the
need to supply detached blob storage when not using serialization is
burdensome.

To address these issues, this change creates a default in memory
detached blob storage which also support serialization and
deserialization, so that detached, serialization and rehydration can be
used in the presence of blobs without the need to pass a custom storage
mechanism which doesn't align with our existing surface area.

For testing I've refactored our existing detached blob storage to
support testing with both a custom detached blob storage specified, and
an undefined detached blob storage which will result in using the new in
memory storage. These test cover attaching, serializing, and
deserializing with blobs.

Lastly IDetachedBlobStorage is deprecated and replaced with a default in
memory store for detached blobs. IDetachedBlobStorage will be removed in
a future release without a replacement. Blobs created while detached
will be stored in memory to align with attached container behavior.

[AB#8049](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/8049)

[AB#5179](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/5179)

---------

Co-authored-by: Tony Murphy <anthonm@microsoft.com>