Add implementation of getKeysInOffsetRange api on Attribution Collection (#21083)
## Description

Add implementation of getKeysInOffsetRange api on Attribution
Collection. This is a requirement from Loop app to integrate the FF
provided attribution.
API: 
/**
* Retrieves all the [Offset, Attribution key] pairs for the provided
offset range. Note:
	 * The returned array is sorted by offset.
* The first offset in response could be lower than the startOffset as
the Attribution Key for the startOffset
* could start at a lower offset than the startOffset in case where
Attribution key offset boundaries don't
	 * align exactly with startOffset.
* Example: If the Attribution Offsets in the segment is [0, 10, 20, 30,
40] and request is for (startOffset: 5, endOffset: 25),
* then result would be [(offset: 0, key: key1), (offset:10, key: key2),
(offset:20, key: key3)].
* @param channel - When specified, gets attribution keys associated with
a particular channel.
* @returns - undefined if the provided channel is not found or list of
attribution keys along with
	 * the corresponding offset start boundary.
	 */
`
	getKeysInOffsetRange(
		startOffset: number,
		endOffset?: number,
		channel?: string,
	): { offset: number; key: AttributionKey }[] | undefined;`
	
This api returns a list of {offset, Attribution key} pairs for the
provided offset range. It helps the app to get attribution keys on a
segment in a single call if they want to fetch multiple for startOffset
till endOffset or from startOffset till the end of segment instead of
making the app to query for key for each offset.

---------

Co-authored-by: Jatin Garg <jatingarg@Jatins-MacBook-Pro-2.local>