fix(core-utils): Fix LazyPromise getter (#21207)
## Description

This PR modifies the getter function for `LazyPromise` to prevent it
from invoking the callback - that is quite an unexpected side effect of
running toString.

We found a case where the events_pkg polyfill was logging the target of
an "addListener" call - happened to be ContainerRuntime - and Node's
console.warn implementation formats the args to print them out. This
eventually cascaded down to calling toString on a LazyPromise which
triggered its callback, but the objects in question had not been
initialized properly so it asserted.