fix: Handle SharedTree events subscriptions with the same listener (#21203)
## Description

This fixes a bug in which the deregister function returned from the
event registration function `on` would deregister multiple events if the
same event listener object was registered multiple times. Each
deregistration function now correctly deregisters only the registration
that produced it, regardless of the object identity of the listener
function. The new `Off` type is added for clarity since it is used as a
map key, and it also helps document the return value for customers.

This also removes an unnecessary event subscription for LazyTree that
was never getting cleaned up (and accidentally relied on the previous
deregister behavior to get deregistered alongside a duplicate
registration).