Fix pnpm lockfile (#21209)
I goofed while hand-editing the pnpm-lock.yaml to remove a
MacOS-specific addition:

On MacOS, the 'path-dirname' sets requireBuild: true.
On Linux, the 'path-dirname' package omits requireBuild.

My hand edit explicitly set 'requireBuild: false', which is probably
benign, but not what you get when the lockfile is generated on Linux.

**Note:** When platforms disagree, we consider Linux to be
authoritative. There are two reasons. The first is that our CI runs on
Linux. The second is that everyone has access to a Linux environment
(WSL, Codespace, Docker, etc.), but not everyone has access to MacOS.