server: dispose socket listeners and log connection metrics (#21211)
## Description

We know there is a memory leak in the Nexus server, and there has been
for a long time. I don't think this will fix it, but it seems like a
good idea to intentionally clean up socket connections on the disconnect
event.

Also, this PR adds a new Metric around socket connections to log:
connection duration, disconnect reason, and transport type (websocket or
long-polling). Disconnect reason is used to determine if the connection
terminated "happily" or not.