Fix typing when setting SharedTree object fields to `undefined` (#21222)
## Description

The typing of content inserted into an object field is currently
incorrect because it does not account for `undefined`. The code in
various layers was typed such that the inserted content would never be
`undefined` when in fact it can be. This PR fixes the typing by changing
several functions to allow for undefined. The functions have been given
overloads such that if the input value is known not to be undefined,
then undefined will not possibly be returned - this is type safe and
removes the need for defined-ness asserts on the output values in some
places.

This PR also has some small cleanups, including adding a missing
`length` property to FlexTreeField and removing an unused helper utility
function.