Client/Server Disentanglement: add gitresources to driver-definitions (#21221)
**Background**
This is a part of a series of changes that remove client -> server
dependencies. The end goal is for the Fluid server/clients to agree on
the wire protocol, but allow compatible divergence of the typings
applied to the deserialized payloads. This work is required to support
the new 2.0 release process (including SemVer) without imposing new
restrictions on the service team. It is also motivated by a consensus
among owners that the code involved is largely stable and that sharing
has more downsides than upsides.

**This Change**
This change adds a handful of types used to strongly type deserialized
REST payloads from the Git service to 'driver-definitions'.

It also renames the types during export to use an 'IGit' prefix to avoid
collision with summary-related types that also describe trees of blobs.
(Previously, these types were imported as '* as git from
"<server_package>"'.)