gitrest: Allow enforcing file size write limits (#19285)
## Description

For reliability, it is important that we can prevent summary uploads to
Redis that are too large to safely handle. This PR adds the
functionality and configuration option to limit `writeFile` size in
RedisFs.

Additionally, we need to protect the service from writing too large of
files. After a file is written in compressed format, it is difficult to
know how large it will be when unpacked. In severe cases, unpacking a
too-compressed summary blob can kill the service by using too much
memory. This PR adds write limits for the default FS as well.