improvement(build-tools): Convert build-cli to ESM (#20972)
Updates the build-cli project to be ESM instead of CJS. For the
oclif-related changes I followed this guide:
https://oclif.io/docs/esm/#migrating-a-cjs-plugin-to-esm

Summary of changes:

- Ran ts2esm to add extensions to source file import statements.
- Upgraded chalk, fs-extra, and node-fetch to more recent versions with
better ESM support.
- jssm's types aren't set up quite right for TypeScript to resolve them
when using Node16/NodeNext module resolution, so I patched in support. I
opened https://github.com/StoneCypher/fsl/issues/1295 to discuss options
for upstreaming the changes.
- I've limited the patched files to package.json, rollup
configs, and the two new rolled up declaration files. Critically, this
means the patch has _no runtime changes_. The only changes are in the
types and config files. This should make it almost impossible for there
to have been runtime bugs introduced by these changes.
- Updated the install build tools step in the CI workflows to use npm
link to make flub globally available. I removed the hack shim we only
used in CI.
- Updated the syncpack config to account for new dependencies.
- Regenerated the packlists. This should be done as a part of the build
but it isn't yet so I updated them opportunistically.
- Renamed the dangerfile to .cts so it's always CommonJS like it used to
be. Seemed the safest approach. If we want it to be ESM we can make that
change separately.