Rename some public event types (#21233)
This renames some of the public types from the SharedTree event library.
The new names have been chosen to be more cohesive as well as to reduce
the change of collision/confusion with any built-in types (e.g.
"Event").

Also, some of the generic typing in the event library has been slightly
simplified.

This is a breaking change - users of the types (e.g. `ISubscribable`)
will need to change to the new name (e.g. `Listenable`).