[Summarization] Refactor summary in progress validation during summarize (#21208)
This is a pre-cursor to
https://github.com/microsoft/FluidFramework/pull/21186 which removes
`trackState`.

In `SummarizerNode::summarize`, removed the validation that a summary in
progress. This validation exists to ensure that `startSummary` was
called before `summarize` so that we can track state for incremental
summaries.
However, this may not be the case in tests where
`ContainerRuntime::summarize` is called directly. Currently, these tests
set `trackState` to false to indicate that it is just interested in the
summary tree without the tracking for incremental summary. The same can
be achieved by setting `fullTree` to true and we can remove the
redundant `trackState` option.

The scenario where this validation is still useful is when `fullTree` is
true in the regular summarization flow via
`ContainerRuntime::submitSummary`. In this scenario, the validation will
still happen in `SummarizerNode::validateSummary` for single-commit
summaries and in `SummarizerNode::completeSummary` for two-commit
summaries. Added unit tests to validate this.