[GC] Added GC performance test for benchmark the performance of initializeOrUpdateGCState (#21269)
Added peformance test to benchmark the peformance of
`initializeOrUpdateGCState` which is known to be expensive when number
of unreferenced nodes are high. This is due to the creation and
initialization of timer for every unreferenced node.

The test added benchmarks the peformance of `initializeOrUpdateGCState`
with 5000, 15000 and 30000 unreferenced nodes. Once the timers issue is
fixed, the performance of these tests should increase.