feat(build-tools): policy for entrypoint linting (#21240)
Policy:
- Package "exports" other than those under "./internal" should be linted
by api-extractor.
- A single "./internal" or alternatively "." export should also be
linted by api-extractor with a bundle config. Recently this was
"check:release-tags" and now is "check:exports:bundle-release-tags".

Unfortunately, api-extractor will only lint a single entrypoint at a
time and needs the config file to be local. So, this means lots of files
and script entries.

Build config:
 - Add "check:exports" to "lint" tasks.

Apply policy to client-utils and container-definitions.

[AB#8141](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/8141)