build: Update API-Extractor config to generate separate reports for each release level (except internal) (#21267)
Each package now generates API reports for each release level alpha,
beta, and public. These reports correspond with the trimmed roll-ups
API-Extractor generates for each release level. This offers convenient
insight into which APIs of a given package are available in which
entrypoint(s). It also offers a mechanism for tracking API items as they
are promoted to higher levels of support guarantees.

For now, we are not generating reports for internal-level exports. This
is intentional but does not have to be a permanent change. If we decide
in the future that it would be valuable to track our internal API
surfaces, we can add that level to the set of reports we generate.

These changes also exclude the packages under `build-tools`, as those
packages don't leverage type-trimming, nor do they manage multiple
entrypoints.


[AB#7523](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/7523)

### Some notes for reviewers

1. Reports for empty package APIs are still generated, and simply don't
contain export details. We may wish to file a feature request to make
this a bit cleaner. Probably append a notice that there are no exports
at the given release level.
2. "Trimmed" reports will contain "untrimmed" imports. This matches the
behavior of the corresponding typings rollups, but it's not particularly
interesting from a report perspective, and could create noise in
automated review enforcement. We should probably file a feature request
around this too, possibly just a way to opt out of including imports in
the report file (I don't think we generally want them there anyways).