[Summarization] Added a document that describes the format of summaries and snapshots (#21264)
Added a document (.md file) to the container runtime's summary folder.
it describes the format of summaries and snapshots. This should help
developers understand how to look at the data, how to make changes, etc.
It should also help users understand the format of their data in summary
/ snapshot.

The documentation goal is to have a README.md in the same folder that
describes overall summarization and it links to this document.