send empty local intervals (#21230)
Currently if a local interval from `clients[0]` is rebased over a
removal from `clients[1]` which removes the entire contents of the
interval, and that interval ends up at one of the extents of the
sequence (the beginning or end), that local interval is removed and
never sent to other clients. If this causes the interval collection to
become empty this creates an inconsistency where only `clients[0]` would
see the collection label in `getIntervalCollectionLabels`.

If `mergeTreeReferencesCanSlideToEndpoint` is enabled, we should treat
intervals at start/end just like any other empty interval and still send
it to all the other clients.