Factor out SharedTree field defaulting into a separate function (#21310)
## Description

Currently, when node data is converted into a MapTree, it has any
missing fields populated by `FieldProvider`s at the same time. This PR
splits the defaulting into a separate step. Now, node data is converted
into MapTrees without populating any defaults. Then, that MapTree can be
augmented by a separate function that adds the defaults to the MapTree.

Also, this PR introduces a distinction between `FieldProvider`s that
need a context (e.g. to generate node identifiers) and those that do not
(e.g. to default to a constant like `undefined`). The defaulting code
will produce both if it is given the necessary context, but if it is not
given any context then it will only generate the defaults that it can
(the non-contextual, constant ones).

This also adds a more comprehensive unit test of field defaulting.

### Why?

An upcoming change will allow MapTrees to be stored in unhydrated nodes
so that those unhydrated nodes can be read before being inserted into
the tree. When the MapTrees are generated for the unhydrated nodes, we
have enough information to populate the "constant" defaults, but we
don't yet have the context for the rest of the defaults - those will
have to wait until the tree is inserted, at which time the defaulting
pass will be run again.