Fixes odsp single commit summaries test (#21288)
Part of
[AB#7929](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/7929)

What is happening in these odsp tests are that summarizers are loading
from older summaries in ODSP due to caching. Because of this, these
summarizers will attempt to summarize and fail and update their cache.
This fixes two of those tests. LatestSummaryRefSeqNum mismatch is
occurring because the summarizer will receive the newer ack and thus
update the running summarizer, but its state is the old state and thus
we have a real mismatch.

This begs a new question, because we will catch this error earlier in
the call stack, we will emit this error when it was originally a generic
event.

There will be follow-up for the next tests
- "Summarizer fetches expected number of times Non-Compat Loading
Summary from older version should fetch"
- "Summarizer fetches expected number of times Non-Compat Second
summarizer from latest should not fetch"