[GC] Remove logic that is not needed after aliasing and tombstone telemetry changes (#21312)
This PR removes the following logic which is not needed anymore:
- Removes the tracking of tombstone nodes in blob manager. This was used
for sending tombstone telemetry which was moved into the garbage
collector.
- Removes the back-compat logic to notify GC when a root data store is
created. With aliasing, a data store is never created as a root. It is
instead created as a regular data store and later aliased at which point
GC will be notified of its referenced.
Note that with this change, if a client running runtime version where
aliasing did not exist creates a root data store, GC will not be
notified. There is no correctness issue because the only reason GC had
to be notified of this reference is to reset unreferenced timestamp, if
any. However, since its a new data store, it won't be referenced (a data
store could not be marked root after creation).