ci: Publish patch GitHub releases automatically (#21309)
Port of #21254 to the RC4 release branch.

Now that the GitHub release process has been working stably for several
months, I've updated the workflow to auto-publish patch releases. This
means that the person doing the release won't need to do anything extra
to get the release published on GitHub.

Applies to patches only; minor and major releases are still left as
drafts.