improvement(tree): Better error messaging when setting / shadowing index properties of array nodes (#21335)
Shadowing of index properties on array nodes leads to questionable
behavior. To simplify our policies here, we will simply forbid it. We
can't easily prevent doing this at the type level, but we can ensure
that users are presented with actionable errors at construction of
invalid array nodes.

From an implementation standpoint, we can accomplish this by simply
throwing an error in the `defineProperty` trap of the array proxy.

Also adds useful error messaging if a user attempts to directly set the
value of an index property.

Note: if/when we allow setting of (valid) index properties in the
future, this implementation strategy will need to change somewhat.


[AB#7753](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/7753)