Copy events set on unhydrated TreeNodes to their hydrated counterparts (#21352)
## Description

Registering an event on an unhydrated node, e.g.:

```ts
const foo = new Foo({...});
Tree.on(foo, "nodeChanged", () => {});
```

is currently pointless, because the event will never fire since the
unhydrated node cannot be mutated.

This PR now copies all the event registrations on that node to the new,
hydrated node after insertion occurs. That way, the events will be
preserved and will fire if the _hydrated_ node is later mutated.