Error when reading identifiers from unhydrated SharedTree nodes (#21366)
## Description

Unhydrated nodes cannot produce "auto-generated" identifiers, because
doing so requires a context containing the FF runtime, which is not
available until after a node is inserted. Therefore, it is impossible
for users to read auto-generated identifiers from unhydrated nodes. This
PR updates the behavior such that if they attempt to do so, we will
return a helpful error message rather than misleadingly returning
`undefined`.