[GC] Make getAttachGCData required on IFluidDataStoreChannel and make implementation consistent (#21347)
`getAttachGCData` was added to `IFluidDataStoreChannel` in 2.0.0-rc.2.0
4 months ago via #19275. This PR does the following:
- Makes `getAttachGCData` required on `IFluidDataStoreChannel`.
- Makes `getAttachGCData` logic consistent across channel collection,
data store context, data store runtime and channel contexts. I found it
confusing to follow the existing flow due to the implementation being
different for various layers. This is how it works now:
- `ChannelCollection` and `FluidDataStoreRuntime` both implements
`getAttachGCData` which iterates through all it's children and gets
their GC data by calling `getAttachGCData`.
- `FluidDataStoreContext` and `LocalChannelContext` both implement
`getAttachGCData` which returns the GC data for this node.
- Decouples `getAttachSummary` and `getAttachGCData` in
`ChannelCollection:generateAttachMessage`.


[AB#8199](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/8199)