supportGetSnapshotApi in setInitialSnapshot (#21313)
This is a prerequisite to periodic snapshot refresh.
Boolean supportGetSnapshotApi is used in serializedStateManager to
determine the api call to storage to retrieve the snapshot. We recently
added a new api getSnapshot to get portion of the snapshot grouped by
id.
We will need to allow making this call for newly created containers in
case we refresh the snapshot after some period of time or after some
number of ops being processed.
This change adds that boolean value after the newly created container is
attached.