fix(devtools): Fixed bug causing multiple focuses on Audience tables (#21303)
[AB#4839](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/4839)

## Description

This PR fixes an a11y bug where tab navigating to the Audience table
would show multiple focuses on the table cell and info icon, without
being able to interact with it to show the tooltip text. The fix ensures
only the info icon is accessible with tab/arrow navigation.

#### Notes
I tried using the built in utilities like useArrowNavigationGroup and
tweaking tabIndex, but that didn't work.
[reference](https://react.fluentui.dev/?path=/docs/components-table--default)
Updating from TableHeaderCell to TableCell seemed to do the trick
without any major changes to the appearance.