Enable single commit summary via supportedFeatures in routerlicious (#21380)
Enable single-commit summaries via `summarizeProtocolTree` policy flag
in `IDocumentServicePolicies` - which for routerlicious will be set when
`enable_single_commit_summary` flag is provided in the
`supportedFeatures` property bag in IConnected. This change is
particularly aimed at enabling slow rollouts of FRS clients on the
single-commit summary flow.

1. updated routerlicious-driver's documentService to check the supported
features and set the `summarizeProtocolTree` driver policy. Container
uses this flag to check whether or not the protocol tree should be added
to the summary before sending the summary.
2. Updated `containerStorageAdapter` to always initialize
`ProtocolTreeStorageService`.
3. `ProtocolTreeStorageService` uses a callback to check whether it
should upload protocoltree with summary or not. This is done to ensure
we are using the most updated value of driver policy which will be set
correctly only when the delta connection is established.
4. Renamed `forceEnableSummarizeProtocolTree` to
`shouldSummarizeProtocolTree`.


[AB#7864](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/7864)

(Also this PR is a recommit of the original commit
here(https://github.com/microsoft/FluidFramework/pull/21060) which had
to be backed out because the tests were failing for odsp. I have made a
fix for that change by ensuring _summarizeProtocolTree property of
containerStorageAdapter is updated to the correct value always. )