Handle location redirection URL error in getSharingInfo call in odsp driver (#21277)
## Description

ADO: https://dev.azure.com/fluidframework/internal/_workitems/edit/8006

After a tenant changes their domain, this leads existing workspaces on
the Loop app to stop getting sharing link to the newly created pages.
This is because after a redirect, the authorization headers are not sent
by the browser. And even if were to just retry with the headers set,
SharePoint requires for the auth tokens to be generated for that
particular domain.

This PR handle causes the browser to not handle the redirect implicitly
by supplying the redirect:manual header in the request and handles the
redirect by itself. It will successfully, return the share link to the
app.

On the load of the same container, we already have
resolveWithLocationRedirectionHandling which handles the redirect during
load and also signals the app about the new domain.

Now, I am using the common redirect header instead of prefer header for
VROOM Apis as getSharingInfo is not a VROOM api. Redirect header works
with all APIs.

---------

Co-authored-by: Jatin Garg <jatingarg@Jatins-MacBook-Pro-2.local>