[Summarization] Removed flaky summarization test for validating data store count in summarize event (#21422)
The following test is flaky - "should generate correct summary stats
with changed and unchanged data stores".
It uses the default heuristics-based summaries which are non-determinic
in terms of when a summary would start so it can take longer than 2
seconds which is the timeout for local server tests.

The test can be fixed by using `summarizeOnDemand`. However, this and
other tests in the file don't add any value. They check that the count
of data store is correct in `summarizeTelemetry`. This telemetry has
been there for 2+ years and doesn't need an end-to-end test to validate
it works.


[AB#6513](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/6513)